Sprint-1 Code shifted in a separate file for Shifat
